InboxRepository: Updating to return count as long instead of int

+++ src/main/java/tech/aroma/banana/data/InboxRepository.java
@@ -52,6 +52,6 @@ public interface InboxRepository
     
     void deleteAllMessagesForUser(@Required String userId) throws TException;
     
-    int countInboxForUser(@Required String userId) throws TException;
+    long countInboxForUser(@Required String userId) throws TException;
     
 }
